munch-api/requirements.txt
requests
flask
"flask[async]"
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
psycopg2-binary
aiohttp